ai job coach/requirements.txt
Flask==3.1.1
Flask_Login==0.6.3
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
PyPDF2==3.0.1
python-dotenv==1.1.1
Requests==2.32.4
Werkzeug==3.1.3
